+++ src/autograd/backward/ExponentsBackward.cpp
@@ -37,7 +37,7 @@ Exp2Backward::Exp2Backward(const Tensor& output) : Node(1), saved_output_(output
 std::vector<Tensor> Exp2Backward::apply(std::vector<Tensor>&& grads) {
     if (grads.empty()) throw std::runtime_error("Exp2Backward: no grads");
     // grad_x = grad_output * y * ln(2)
-    return {grads[0] * saved_output_ * static_cast<float>(M_LN2)};
+    return {grads[0] * saved_output_ * M_LN2};
 }
 
 // ===================================
@@ -48,7 +48,7 @@ Log2Backward::Log2Backward(const Tensor& input) : Node(1), saved_input_(input) {
 std::vector<Tensor> Log2Backward::apply(std::vector<Tensor>&& grads) {
     if (grads.empty()) throw std::runtime_error("Log2Backward: no grads");
     // grad_x = grad_output / (x * ln(2))
-    return {grads[0] / (saved_input_ * static_cast<float>(M_LN2))};
+    return {grads[0] / (saved_input_ * M_LN2)};
 }
 
 // ===================================
@@ -59,7 +59,7 @@ Log10Backward::Log10Backward(const Tensor& input) : Node(1), saved_input_(input)
 std::vector<Tensor> Log10Backward::apply(std::vector<Tensor>&& grads) {
     if (grads.empty()) throw std::runtime_error("Log10Backward: no grads");
     // grad_x = grad_output / (x * ln(10))
-    return {grads[0] / (saved_input_ * static_cast<float>(M_LN10))};
+    return {grads[0] / (saved_input_ * M_LN10)};
 }
 
 } // namespace autograd
